Unsquashfs: Fix warning if time_t is 64-bits on 32-bit system
Traditionally time_t is a "long int" which matches "%ld". On 32-bit
systems both are 32-bit, and on 64-bit systems both are 64-bit.

However there are recent efforts to make time_t 64-bit on 32-bit systems.
For example recent versions of glibc support 64-bit time_t on 32-bit systems
if _TIME_BITS=64.

This breaks the previous fact that time_t always matches "%ld".

Fix this by moving to "%lld" and casting inode->time to (long long int)
to ensure it matches.

Fixes issue https://github.com/plougher/squashfs-tools/issues/260

+++ squashfs-tools/unsquashfs.c
@@ -3596,7 +3596,7 @@ void pseudo_print(char *pathname, struct inode *inode, char *link, long long off
 	else if(res >= 12)
 		EXIT_UNSQUASH("snprintf returned more than 11 digits in pseudo_print()\n");
 
-	res = dprintf(writer_fd, "%s %c %ld %o %s %s", filename, type, inode->time, inode->mode & ~S_IFMT, userstr, groupstr);
+	res = dprintf(writer_fd, "%s %c %lld %o %s %s", filename, type, (long long int) inode->time, inode->mode & ~S_IFMT, userstr, groupstr);
 	if(res == -1)
 		EXIT_UNSQUASH("Failed to write to pseudo output file\n");
